new version 0.3.0, fixing NumerApiManager
